upload-service/requirements.txt
Flask==2.0.3
requests==2.26.0
Werkzeug==2.0.3
mysql-connector-python

